Deduplicate ParameterInfoTests
