+++ src/System.Reflection/tests/ParameterInfoTests.cs
@@ -5,6 +5,8 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Runtime.InteropServices;
+using System.Runtime.CompilerServices;
+using System.Runtime.Serialization;
 using Xunit;
 
 namespace System.Reflection.Tests
@@ -96,6 +98,33 @@ namespace System.Reflection.Tests
         {
             ParameterInfo parameterInfo = GetParameterInfo(type, name, index);
             Assert.NotNull(parameterInfo.RawDefaultValue);
+        }
+
+        [Fact]
+        public void RawDefaultValue_Enum()
+        {
+            ParameterInfo p = GetParameterInfo(typeof(ParameterInfoMetadata), "Foo1", 0);
+            object raw = p.RawDefaultValue;
+            Assert.Equal(typeof(int), raw.GetType());
+            Assert.Equal<int>((int)raw, (int)BindingFlags.DeclaredOnly);
+        }
+
+        [Fact]
+        public void RawDefaultValueFromAttribute()
+        {
+            ParameterInfo p = GetParameterInfo(typeof(ParameterInfoMetadata), "Foo2", 0);
+            object raw = p.RawDefaultValue;    
+            Assert.Equal(typeof(int), raw.GetType());
+            Assert.Equal<int>((int)raw, (int)BindingFlags.IgnoreCase);
+        }
+
+        [Fact]
+        public void RawDefaultValue_MetadataTrumpsAttribute()
+        {
+            ParameterInfo p = GetParameterInfo(typeof(ParameterInfoMetadata), "Foo3", 0);
+            object raw = p.RawDefaultValue;    
+            Assert.Equal(typeof(int), raw.GetType());
+            Assert.Equal<int>((int)raw, (int)BindingFlags.FlattenHierarchy);
         }
 
         [Theory]
@@ -193,6 +222,27 @@ namespace System.Reflection.Tests
             Assert.True(prov.IsDefined(attrType, true));
         }
 
+        [Fact]
+        public void VerifyGetCustomAttributesData()
+        {
+            ParameterInfo p = GetParameterInfo(typeof(ParameterInfoMetadata), "MethodWithCustomAttribute", 0);
+            foreach (CustomAttributeData cad in p.GetCustomAttributesData())
+            {
+                if(cad.AttributeType == typeof(MyAttribute))
+                {
+                    ConstructorInfo c = cad.Constructor;
+                    Assert.False(c.IsStatic);
+                    Assert.False(c.IsPublic);
+                    ParameterInfo[] paramInfo = c.GetParameters();
+                    Assert.Equal(1, paramInfo.Length);
+                    Assert.Equal(typeof(int), paramInfo[0].ParameterType);
+                    return;
+                }
+            }
+
+            Assert.True(false, "Expected to find MyAttribute");
+        }        
+
         public static IEnumerable<object[]> s_CustomAttributesTestData
         {
             get
@@ -231,6 +281,47 @@ namespace System.Reflection.Tests
             Assert.Equal(expected, parameterInfo.GetRequiredCustomModifiers());
         }
 
+        [Theory]
+        [MemberData(nameof(VerifyParameterInfoGetRealObjectWorks_TestData))]
+        public void VerifyParameterInfoGetRealObjectWorks(MemberInfo pretendMember, int pretendPosition, string expectedParameterName)
+        {
+            // Regression test for https://github.com/dotnet/corefx/issues/20574
+            //
+            // It's easy to forget that ParameterInfo's and runtime-implemented ParameterInfo's are different objects and just because the
+            // latter doesn't support serialization doesn't mean other providers won't either. 
+            //
+            // For historical reasons, ParameterInfo contains some serialization support that subtypes can optionally hang off. This
+            // test ensures that support doesn't get vaporized.
+
+            // Just pretend that we're BinaryFormatter and are deserializing a Parameter...
+            IObjectReference podParameter = new PodPersonParameterInfo(pretendMember, pretendPosition);
+            StreamingContext sc = new StreamingContext(StreamingContextStates.Clone);
+            ParameterInfo result = (ParameterInfo)(podParameter.GetRealObject(sc));
+
+            Assert.Equal(pretendPosition, result.Position);
+            Assert.Equal(expectedParameterName, result.Name);
+            Assert.Equal(pretendMember.Name, result.Member.Name);
+        }
+
+        public static IEnumerable<object[]> VerifyParameterInfoGetRealObjectWorks_TestData
+        {
+            get
+            {
+                Type t = typeof(PretendParent);
+                ConstructorInfo ctor = t.GetConstructor(new Type[] { typeof(int), typeof(int) });
+                MethodInfo method = t.GetMethod(nameof(PretendParent.PretendMethod));
+                PropertyInfo property = t.GetProperty("Item");
+
+                yield return new object[] { ctor, 0, "a" };
+                yield return new object[] { ctor, 1, "b" };
+                yield return new object[] { method, -1, null };
+                yield return new object[] { method, 0, "x" };
+                yield return new object[] { method, 1, "y" };
+                yield return new object[] { property, 0, "index1" };
+                yield return new object[] { property, 1, "index2" };
+            }
+        }
+
         private static ParameterInfo GetParameterInfo(Type type, string name, int index)
         {
             ParameterInfo[] parameters = GetMethod(type, name).GetParameters();
@@ -245,6 +336,12 @@ namespace System.Reflection.Tests
         // Metadata for reflection
         public class ParameterInfoMetadata
         {
+            public void Foo1(BindingFlags bf = BindingFlags.DeclaredOnly) { }
+            public void Foo2([CustomBindingFlags(Value = BindingFlags.IgnoreCase)] BindingFlags bf) { }
+            public void Foo3([CustomBindingFlags(Value = BindingFlags.DeclaredOnly)] BindingFlags bf = BindingFlags.FlattenHierarchy ) { }
+
+            public void MethodWithCustomAttribute([My(2)]string str, int iValue, long lValue) { }
+
             public void Method1(string str, int iValue, long lValue) { }
             public void Method2() { }
             public void MethodWithArray(string[] strArray) { }
@@ -274,5 +371,39 @@ namespace System.Reflection.Tests
             public void GenericMethod(T t) { }
             public string GenericMethodWithDefault(int i, T t = default(T)) { return "somestring"; }
         }
+
+        private class MyAttribute : Attribute
+        {
+            internal MyAttribute(int i) { }
+        }
+
+        internal sealed class CustomBindingFlagsAttribute : UsableCustomConstantAttribute
+        {
+            public new object Value { get { return RealValue; } set { RealValue = value; } }
+        }
+
+        internal abstract class UsableCustomConstantAttribute : CustomConstantAttribute
+        {
+            public sealed override object Value => RealValue;
+            protected object RealValue { get; set; }
+        }
+
+        private sealed class PretendParent
+        {
+            public PretendParent(int a, int b) { }
+            public void PretendMethod(int x, int y) { }
+            public int this[int index1, int index2] { get { throw null; } }
+        }
+
+        private sealed class PodPersonParameterInfo : MockParameterInfo
+        {
+            public PodPersonParameterInfo(MemberInfo pretendMember, int pretendPosition)
+            {
+                // Serialization can recreate a ParameterInfo from just these two pieces of data. Of course, this is just a test and no one 
+                // ever told this Member that it was adopting a counterfeit Parameter, but this is just a test...
+                MemberImpl = pretendMember;
+                PositionImpl = pretendPosition;
+            }
+        }        
     }
 }
